Fix tag icons and warning things.
Not very sure what is happening and I don't like it when that happens.
I want to understand.

Things I do not understand: the font name change, the neccesity of the
font-weight attribute, the fact that that attribute NEEDS to be 900?
Will ask on the issue that started this, as this entire thing has been
because another user had the same problem, will perhaps ask later.